AIIncidentIntelligencePlatform/event-processor/requirements.txt
confluent-kafka==2.3.0
elasticsearch==8.11.0
redis[hiredis]==5.0.1
python-dotenv==1.0.0
prometheus-client==0.19.0